+++ app/src/main/java/com/ilyaeremin/graphicmodule/IntervalListener.java
@@ -1,7 +1,7 @@
 package com.ilyaeremin.graphicmodule;
 
-import android.support.annotation.IntRange;
+import android.support.annotation.FloatRange;
 
 public interface IntervalListener {
-    void onChanged(@IntRange(from = 0, to = 100) int leftBound, @IntRange(from = 0, to = 100) int rightBound);
+    void onChanged(@FloatRange(from = 0, to = 100) float leftBound, @FloatRange(from = 0, to = 100) float rightBound);
 }
